Refactor: GridCombatSystem split
Introduced new class CombatGridManager responsible for managing grid. Moved there all grid-related methods from Combat System class
