+++ Assets/Scripts/CombatGridManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce985f57d80cbf14884a07cd151ef5be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
